+++ .gitignore
@@ -62,3 +62,4 @@ dkms.conf
 
 # macOS
 .DS_Store
+.vscode
