+++ concurrent.c
@@ -1,7 +1,9 @@
 #define _GNU_SOURCE
 #include "project.h"
 #include "utils.h"
+#include "affinity.h"
 #include "concurrent.h"
+#include "tuples.h"  // In case tuple_t definition is needed
 #include <pthread.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -12,7 +14,7 @@ typedef struct {
     int tuples_index;
     int tuples_length;
     int partition_count;
-    // In concurrent mode, all threads share the same global partitions.
+    // Shared partitions.
     tuple_t **partitions;
     int *partition_indexes;
     pthread_mutex_t *partition_mutexes;
@@ -23,6 +25,10 @@ void *write_to_partitions(void *void_args) {
     if (!void_args)
         return NULL;
     thread_args_t *args = (thread_args_t *)void_args;
+    
+    // Set affinity if enabled.
+    set_affinity(args->thread_id);
+    
     if (!args->tuples || !args->partitions || !args->partition_indexes || !args->partition_mutexes)
         return NULL;
     double start = get_time_in_seconds();
@@ -35,7 +41,6 @@ void *write_to_partitions(void *void_args) {
     }
     double end = get_time_in_seconds();
     args->thread_time = end - start;
-    //printf("Concurrent thread %d finished\n", args->thread_id);
     return NULL;
 }
 
@@ -45,7 +50,6 @@ int run_concurrent_timed(tuple_t *tuples, int tuple_count, int thread_count, int
     if (!tuples)
         return -1;
     
-    // Compute effective capacity for this run.
     int effective_capacity = (tuple_count / partition_count) * PARTITION_MULTIPLIER;
     if (effective_capacity > global_capacity)
         effective_capacity = global_capacity;
@@ -55,12 +59,10 @@ int run_concurrent_timed(tuple_t *tuples, int tuple_count, int thread_count, int
     int base_segment_size = tuple_count / thread_count;
     double overall_throughput = 0.0;
     
-    // Reset the indexes for the partitions used (only current partition_count).
     for (int i = 0; i < partition_count; i++) {
         global_partition_indexes[i] = 0;
     }
     
-    // Allocate and initialize mutexes for each partition.
     pthread_mutex_t *mutexes = malloc(partition_count * sizeof(pthread_mutex_t));
     if (!mutexes) {
         fprintf(stderr, "Failed to allocate mutexes.\n");
@@ -77,22 +79,20 @@ int run_concurrent_timed(tuple_t *tuples, int tuple_count, int thread_count, int
         }
     }
     
-    // Create threads, passing the shared mutexes array.
     for (int i = 0; i < thread_count; i++) {
         int start_index = base_segment_size * i;
         int end_index = (i == thread_count - 1) ? tuple_count : (start_index + base_segment_size);
-        args[i].thread_id = i + 1;
+        args[i].thread_id = i + 1;  // Correct assignment: set thread_id to (i+1)
         args[i].tuples = tuples;
         args[i].tuples_index = start_index;
         args[i].tuples_length = end_index;
         args[i].partition_count = partition_count;
-        args[i].partitions = global_partition_buffers; // Shared by all threads.
+        args[i].partitions = global_partition_buffers;     // Shared among threads.
         args[i].partition_indexes = global_partition_indexes; // Shared.
-        args[i].partition_mutexes = mutexes; // Shared mutexes.
+        args[i].partition_mutexes = mutexes;                // Shared mutexes.
         
         if (pthread_create(&threads[i], NULL, write_to_partitions, &args[i]) != 0) {
             fprintf(stderr, "Concurrent thread creation failed for thread %d\n", args[i].thread_id);
-            // Cleanup mutexes before returning.
             for (int j = 0; j < partition_count; j++) {
                 pthread_mutex_destroy(&mutexes[j]);
             }
@@ -101,7 +101,6 @@ int run_concurrent_timed(tuple_t *tuples, int tuple_count, int thread_count, int
         }
     }
     
-    // Wait for threads to finish and accumulate throughput.
     for (int i = 0; i < thread_count; i++) {
         pthread_join(threads[i], NULL);
         int seg = args[i].tuples_length - args[i].tuples_index;
@@ -110,7 +109,6 @@ int run_concurrent_timed(tuple_t *tuples, int tuple_count, int thread_count, int
     }
     *throughput = overall_throughput;
     
-    // Destroy mutexes and free the mutex array.
     for (int i = 0; i < partition_count; i++) {
         pthread_mutex_destroy(&mutexes[i]);
     }
