+++ readme.md
@@ -1,0 +1,5 @@
+For running the experiemnts run the following two commands:
+```bash
+make all
+make run_all
+```
