+++ tuples.h
@@ -1,0 +1,9 @@
+#ifndef TUPLES_H
+#define TUPLES_H
+
+#include "project.h"
+
+// Declaration of generate_tuples.
+tuple_t *generate_tuples(int count);
+
+#endif
