+++ utils.h
@@ -1,2 +1,10 @@
+#ifndef UTILS_H
+#define UTILS_H
+
+#include <stdint.h>
+#include "project.h"
+
 int hash_to_partition(const unsigned char *key, int partition_count);
 double get_time_in_seconds();
+
+#endif
